requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-extensions==3.2.1
django-filter==22.1
django-pandas==0.6.6
django-phonenumber-field==7.0.1
folium==0.13.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
pandas==1.5.2
phonenumbers==8.13.2
Pillow==9.4.0
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
